refactor: rename package - navigation after authentication success - remove comment
