+++ app/src/main/java/com/pwhs/quickmem/presentation/app/settings/SettingsScreen.kt
@@ -67,6 +67,7 @@ import com.ramcosta.composedestinations.generated.destinations.ChangePasswordSet
 import com.ramcosta.composedestinations.generated.destinations.ProfileScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.UpdateEmailSettingScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.UpdateFullNameSettingScreenDestination
+import com.ramcosta.composedestinations.generated.destinations.UpdateUsernameSettingScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.WelcomeScreenDestination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.result.NavResult
@@ -163,7 +164,12 @@ fun SettingsScreen(
                 }
 
                 SettingUiEvent.NavigateToChangeUsername -> {
-                    // TODO()
+                    navigator.navigate(
+                        UpdateUsernameSettingScreenDestination(
+                            userId = uiState.userId,
+                            username = uiState.username
+                        )
+                    )
                 }
             }
